+++ core/src/main/java/de/mario/roguelette/render/bet/CornerRegion.java
@@ -17,7 +17,7 @@ public class CornerRegion extends InsideBetRegion {
     }
 
     @Override
-    public Bet createBet(int amount) {
+    public Bet createBet(long amount) {
         return new Bet(new CornerBet(firstNumber), amount);
     }
