+++ core/src/main/java/de/mario/roguelette/render/bet/VerticalSplitRegion.java
@@ -17,7 +17,7 @@ public class VerticalSplitRegion extends InsideBetRegion {
     }
 
     @Override
-    public Bet createBet(int amount) {
+    public Bet createBet(long amount) {
         return new Bet(new VerticalSplitBet(firstNumber), amount);
     }
